update: move and remove modules
