+++ app/(main)/index.tsx
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from 'react';
 import {useWindowDimensions, View} from 'react-native';
 import Conversations from '@/src/components/conversations/Conversations';
 import ChatArea from '@/src/components/chat/ChatArea';
-import Info from '@/src/components/chat/Info';
+import Info from '@/src/components/info/Info';
 import {Conversation} from '@/src/models/Conversation';
 import { useNavigation } from 'expo-router';
 import { useTabBar } from '@/src/contexts/tabbar/TabBarContext';
