+++ src/components/chat/message/VoteMessageContent.tsx
