+++ src/components/info/Info.tsx
@@ -2,13 +2,13 @@ import React, {useEffect, useState} from 'react';
 import {ScrollView, Text, TouchableOpacity, View} from 'react-native';
 import {Conversation} from '@/src/models/Conversation';
 import Search from './Search';
-import HeaderInfo from '../info/HeaderInfo';
-import ProfileInfo from '../info/ProfileInfo';
-import ActionsInfo from '../info/ActionsInfo';
-import MediaInfo from '../info/MediaInfo';
-import FilesInfo from '../info/FilesInfo';
+import HeaderInfo from './HeaderInfo';
+import ProfileInfo from './ProfileInfo';
+import ActionsInfo from './ActionsInfo';
+import MediaInfo from './MediaInfo';
+import FilesInfo from './FilesInfo';
 import {Ionicons} from '@expo/vector-icons';
-import GroupInfo from '../info/GroupInfo';
+import GroupInfo from './GroupInfo';
 import {ConversationService} from '@/src/api/services/ConversationService';
 import {useUser} from '@/src/contexts/user/UserContext';
